added new methods & clean up
New method in stock manager, involving selling multiple products if the products have an amount lower then quantity.
Also another new method to select an amount if id is matched & the amount within the product is 0...

within class product, there is a sell number method to sell that individual product a certain number, also if quantity is more then 0.

Lastly there should be a Testing file for App04